use portable format for (u)int64_t types

+++ inc_internal/zt_internal.h
@@ -91,8 +91,8 @@ typedef struct ziti_channel {
     uint64_t last_read;
     uint64_t last_write;
     uint64_t last_write_delay;
-    uint64_t out_q;
-    uint64_t out_q_bytes;
+    size_t out_q;
+    size_t out_q_bytes;
 
     ch_state state;
     uint32_t reconnect_count;
